refactor: pull module loading into prebake
